# design/api/openapi.yaml
openapi: 3.0.3
info:
  title: IDaaS系统API接口文档
  description: |
    IDaaS（Identity as a Service）系统API接口规范，提供统一的身份认证、授权管理和身份治理服务。

    ## 系统特性
    - 统一身份认证与单点登录
    - 多因子认证支持
    - RBAC/ABAC混合权限模型
    - 组织架构管理
    - 应用集成与联邦身份
    - 身份治理与合规审计

    ## 认证方式
    - Bearer Token (JWT)
    - OAuth2.0
    - OpenID Connect

    ## 错误响应格式
    ```json
    {
      "code": "ERROR_CODE",
      "message": "错误描述信息",
      "timestamp": "2024-01-01T12:00:00Z",
      "path": "/api/resource"
    }
    ```
  version: 1.0.0
  contact:
    name: Qoobot开发团队
    email: dev@qoobot.com
    url: https://github.com/qoobot-com/openidaas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.idaas.example.com
    description: 生产环境
  - url: https://test-api.idaas.example.com
    description: 测试环境
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/users/{id}:
    get:
      tags: [users]
      summary: 获取用户详情
      description: 根据用户ID获取用户详细信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: 用户不存在

    put:
      tags: [users]
      summary: 更新用户信息
      description: 更新用户基本信息和档案信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
        '404':
          description: 用户不存在

    delete:
      tags: [users]
      summary: 删除用户
      description: 逻辑删除用户账户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '404':
          description: 用户不存在
...
  /api/users/{id}/roles:
    post:
      tags: [users]
      summary: 分配用户角色
      description: 为用户分配角色和权限
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: 分配成功
        '400':
          description: 请求参数错误
        '404':
          description: 用户或角色不存在
        '409':
          description: 角色已分配

  # 组织架构接口
...
  /api/roles:
    get:
      tags: [roles]
      summary: 查询角色列表
      description: 分页查询角色信息
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: roleName
          in: query
          description: 角色名称模糊查询
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseRole'

    post:
      tags: [roles]
      summary: 创建角色
      description: 创建新的角色定义
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: 角色创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
  /api/roles/{id}/permissions:
    get:
      tags: [roles]
      summary: 获取角色权限
      description: 获取指定角色拥有的权限列表
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 角色ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'

    post:
      tags: [roles]
      summary: 分配角色权限
      description: 为角色分配权限
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 角色ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionRequest'
      responses:
        '200':
          description: 权限分配成功

  # 应用管理接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      description: 页码（从0开始）
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: 每页大小
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    UserStatus:
      type: string
      description: 用户状态枚举
      enum: [ACTIVE, LOCKED, DISABLED, DELETED]
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID
        username:
          type: string
          description: 用户名
        email:
          type: string
          description: 邮箱地址
        mobile:
          type: string
          description: 手机号码
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/UserProfile'
            departments:
              type: array
              items:
                $ref: '#/components/schemas/UserDepartment'
            roles:
              type: array
              items:
                $ref: '#/components/schemas/UserRoleAssignment'
...
    UserProfile:
      type: object
      properties:
        fullName:
          type: string
          description: 真实姓名
        nickname:
          type: string
          description: 昵称
        gender:
          type: integer
          description: 性别（1-男，2-女，0-未知）
        birthDate:
          type: string
          format: date
          description: 出生日期
        avatarUrl:
          type: string
          description: 头像URL
        employeeId:
          type: string
          description: 员工编号
    UserDepartment:
      type: object
      properties:
        deptId:
          type: integer
          format: int64
        deptName:
          type: string
        isPrimary:
          type: boolean
        positionId:
          type: integer
          format: int64
        positionName:
          type: string
    UserRoleAssignment:
      type: object
      properties:
        roleId:
          type: integer
          format: int64
        roleName:
          type: string
        scopeType:
          type: string
          description: 作用域类型
        scopeId:
          type: integer
          format: int64
        expireTime:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 128
          description: 邮箱地址
        mobile:
          type: string
          pattern: '^1[3-9]\d{9}$'
          description: 手机号码
        profile:
          $ref: '#/components/schemas/UserProfile'
        status:
          $ref: '#/components/schemas/UserStatus'

    # 用户部门分配相关
...
    AssignRoleRequest:
      type: object
      required:
        - roleId
      properties:
        roleId:
          type: integer
          format: int64
          description: 角色ID
          minimum: 1
        scopeType:
          type: integer
          description: 作用域类型（1-全局，2-部门，3-项目）
          default: 1
          enum: [1, 2, 3]
        scopeId:
          type: integer
          format: int64
          description: 作用域ID
        expireTime:
          type: string
          format: date-time
          description: 过期时间
        grantReason:
          type: string
          maxLength: 500
          description: 授权原因

    # 部门管理相关
...
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        roleCode:
          type: string
        roleName:
          type: string
        roleType:
          type: integer
          description: 角色类型（1-系统角色，2-自定义角色）
        parentId:
          type: integer
          format: int64
        description:
          type: string
        sortOrder:
          type: integer
          description: 排序
    Permission:
      type: object
      properties:
        id:
          type: integer
          format: int64
        permCode:
          type: string
        permName:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        action:
          type: string
    CreateRoleRequest:
      type: object
      required:
        - roleCode
        - roleName
      properties:
        roleCode:
          type: string
        roleName:
          type: string
        roleType:
          type: integer
          default: 2
        parentId:
          type: integer
          format: int64
        description:
          type: string
        sortOrder:
          type: integer
          description: 排序
          default: 0
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
    AssignPermissionRequest:
      type: object
      required:
        - permissionIds
      properties:
        permissionIds:
          type: array
          items:
            type: integer
            format: int64

    # 应用管理相关
...
    PageResponseRole:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...